requirements.txt
fastapi
uvicorn
pydantic
nltk
requests
pytest
python-dotenv
pytest-asyncio
httpx
pytest-cov